moved HandlerUtils to handler module + added new method to add audit fields wherever needed.
